requirements.txt
python-dotenv
faiss-cpu
tiktoken
langchain-community
pypdf
PyPDF2
langchain_openai
pandas
pinecone
langchain_pinecone
sentence-transformers
streamlit
flask